Shell concept map and presentation workflow
I have included two files.  The .svg is our concept map of the commands we will teach in our session.  The .jpg is our presentation workflow.  The concept map is an overview of all of the commands and what they are used to do.  We did not include any details on how each command works.